backend/requirements.txt
# FastAPI y dependencias core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Email
aiosmtplib==3.0.1

# HTTP requests
aiohttp==3.9.1
httpx==0.25.2

# Reportes PDF
reportlab==4.0.7
PyPDF2==3.0.1

# Reportes Excel/CSV
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9

# Utilidades
python-dotenv==1.0.0
pytz==2023.3

# Desarrollo y testing (opcional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
